% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{checkCalibration_uniCont}
\alias{checkCalibration_uniCont}
\title{Build calibration plots for a \code{uniCont} object}
\usage{
checkCalibration_uniCont(fitted_model, resample_indexes, number_bins,
  weight_lower_bound = NULL, x_axis = "predicted")
}
\description{
Build calibration plots for a \code{uniCont} object
}

