% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{checkCalibrationBase}
\alias{checkCalibrationBase}
\title{Build calibration plots for univariate continuous variables.}
\usage{
checkCalibrationBase(x, y, number_bins, x_labs, y_labs, weights = NULL,
  smooth_order = 3)
}
\arguments{
\item{x}{Continuous variable to be plotted on the x-axis.}

\item{y}{Continuous variable to be plotted on the y-axis.}

\item{number_bins}{Number of bins to use for plotting. Each
bin will be represented by a dot, which is computed by an
average of the points contained in each bin.}

\item{x_labs}{String to be use as label for the x-axis.}

\item{y_labs}{String to be use as label for the x-axis.}

\item{weights}{Numerical vector of the same length as in
\code{x} and \code{y}. To be used in case each point (x,y)
have a weight attached to it. Default is NULL for no weights.}

\item{smooth_order}{Degrees of freedom to generate a natural cubic splines.
 Default to 3. If set to 1 recover a linear trend.}
}
\value{
List with three elements: 1) \code{calibration_objects}
 is a data.frame containing one row for each bin with the x- and y-
 axis coordinates used in the plotting. 2) \code{plot_calibration_points} is
 a ggplot2 object representing the dot plot of the bin points.
 3) \code{plot_smooth_calibration} is a ggplot2 object which contains
 an smooth estimate of the fit, in addition to the dot plot of the bin points.
}
\description{
Build calibration plots for univariate continuous variables.
}

