% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{computeNegativeSquareError}
\alias{computeNegativeSquareError}
\title{Compute the negative of the square error of predictions}
\usage{
computeNegativeSquareError(prediction, target, weight = NULL)
}
\arguments{
\item{prediction}{Numeric vector with univariate predictions.}

\item{target}{Numeric vector with the observed values of the target variables.}

\item{weight}{Numeric vector with the weight given to each prediction point.
 Default is NULL.}
}
\description{
Compute the negative of the square error of predictions. The negative part
is to ensure that the higher the metric is the better.
}

