% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{computeUniContScore}
\alias{computeUniContScore}
\title{Compute scores from univariate and continuous predictions.}
\usage{
computeUniContScore(prediction, target, type, weight = NULL, ...)
}
\arguments{
\item{prediction}{Numeric vector with univariate predictions.}

\item{target}{Numeric vector with the observed values of the target variables.}

\item{type}{Which score function to use. See details.}

\item{weight}{Numeric vector with the weight given to each prediction point.
Default is NULL.}

\item{...}{extra parameters used in a specific score function. Currently not in use.}
}
\value{
Numeric vector with length equal to the number of \code{prediction}
points containing the scores for each prediction point.
}
\description{
Compute scores from univariate and continuous predictions.
}
\details{
Given a set of \code{prediction} and \code{target} values it will
  compute the score function for each prediction point. Possibly each prediction
  point can be associated with a \code{weight}
}

