% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluateProbClass.uniContValidation}
\alias{evaluateProbClass.uniContValidation}
\title{Evaluate univariate continuous predictions over a validation set for a given score function.}
\usage{
\method{evaluateProbClass}{uniContValidation}(pred_obj, resample_indexes, type,
  ignore_tag = FALSE, replicate_index = NULL, ...)
}
\arguments{
\item{pred_obj}{A \code{uniContValidation} object, usually generated
from the \code{datasetResample} object with
some set of predictive models through an interface function to
the \code{tgmUniCont} package.}

\item{resample_indexes}{A \code{datasetResample} object, usually
generated by the \code{generateTestIndexes} function.}

\item{type}{Which type of score function should be used.
See \code{?computeUniContScore} for the updated list of score
functions available.}

\item{ignore_tag}{Before computing scores, we check to see if
 \code{pred_obj} was indeed generated from the \code{resample_indexes}.
 In case of a mismatch an error is issued, unless \code{ignore_tag = FALSE},
 in which case just a warning is issued.}
}
\value{
An S3 object of class \code{multiClassValidationScores}.
}
\description{
Evaluate univariate continuous predictions over a validation set for a given score function.
}
\details{
Given a \code{datasetResample} object and a compatible
\code{uniContValidation} prediction object it computes prediction scores
of some \code{type} for each model used in the validation step.
}

