% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.R
\name{mcGet.uniContValidation}
\alias{mcGet.uniContValidation}
\title{Extract info from \code{uniContValidation} objects}
\usage{
\method{mcGet}{uniContValidation}(x, attr, i = NULL)
}
\description{
Extract info from \code{uniContValidation} objects
}

