% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.R
\name{summarizeValidation.uniCont}
\alias{summarizeValidation.uniCont}
\title{Summarize results from the validation step}
\usage{
\method{summarizeValidation}{uniCont}(fitted_model, ...)
}
\description{
Summarize results from the validation step
}

