% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniCont.R
\name{uniCont}
\alias{uniCont}
\title{Constructor for \code{uniCont} objects}
\usage{
uniCont(resample_indexes, pred, summary_validation = NULL,
  metric_name = NULL, weight_name = NULL)
}
\arguments{
\item{resample_indexes}{S3 object of class \code{datasetResample} with information
about the data used using for training and predictions. Usually
generated by the \code{generateTestIndexes} function.}

\item{pred}{List where each element contains a numeric vector with predictions for
a given test set.}

\item{summary_validation}{Object of class \code{summaryValidation}, usually
created by \code{SummaryValidationMetric} function. Default to NULL, when
no validation step is necessary.}

\item{metric_name}{Character vector indicating the name of the metric used
in the validation step. Default to NULL, when no validation step
is necessary. See the help of \code{evaluateProbClass} function to
which metrics are currently available.}

\item{weight_name}{Character element with the name of the variable used as weight in the
 model training}
}
\value{
S3 object of class \code{uniCont}.
}
\description{
\code{uniCont} objects hold final test set univariate continuous predictions performed by models.
}
\details{
\code{uniCont} objects hold univariate continuous predictions performed by a model over a number
of test sets generated according to some resampling of the data. A tag exist
to connect the predictions with the object that contains information about
the selected test sets. This allows results from different models to be
compared since two \code{uniCont} objects with the same tag are guaranteed
to be generated from the same test set points.
}
\seealso{
\code{\link{generateTestIndexes}},
 \code{\link{summaryValidation}},
 \code{\link{evaluateProbClass}}

Other tgmUniCont.constructors: \code{\link{uniContValidation}}
}

