% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.R
\name{uniContValidation}
\alias{uniContValidation}
\title{Constructor for \code{uniContValidation} objects}
\usage{
uniContValidation(resample_indexes, pred, replicate_index, tune_grid,
  weight_name = NULL)
}
\arguments{
\item{resample_indexes}{S3 object of class \code{datasetResample} with information
about Class Labels and the data used using for training and predictions. Usually
generated by the \code{generateTestIndexes} function.}

\item{pred}{List where each element contains a numeric vector with predictions performed
by a model on the same validation set.}

\item{replicate_index}{Integer number that indicates which validation set was used.}

\item{tune_grid}{It is a matrix with named columns containing tuning parameters of the
model. Each row of the matrix represents tuning parameter values of a specific model.}

\item{weight_name}{Character element with the name of the variable used as weight in the
 model training}
}
\value{
S3 object of class \code{uniContValidation}.
}
\description{
\code{uniContValidation} objects hold validation set predictions performed by different models.
}
\details{
\code{uniContValidation} objects hold univariate continuous predictions performed
by different models over the same validation set. A tag exist
to connect the predictions with the object that contains information about
the selected validation set. In addition the specific replicate index of the
validation set used is stored because several validation sets are contained in the
tagged object.
}
\seealso{
\code{\link{generateTestIndexes}}

Other tgmUniCont.constructors: \code{\link{uniCont}}
}

